requirements.txt
groq==0.9.0
httpx==0.27.0
langchain-core==0.2.0
python-dotenv==1.0.0
streamlit==1.38.0
streamlit-extras==0.4.0
scikit-learn==1.5.0